+++ osa7/blogilista/blogilista-frontend/src/components/BlogList.jsx
@@ -1,43 +1,20 @@
 import { useSelector, useDispatch } from 'react-redux'
+import { Routes, Route, useMatch, Link } from 'react-router-dom'
 import Blog from './Blog'
 import { likeBlog, removeBlog } from '../reducers/blogReducer'
 import { newNotification } from '../reducers/notificationReducer'
 
 const BlogList = ({ user }) => {
-  const dispatch = useDispatch()
   const blogs = useSelector((state) => state.blogs)
-
-  const addLike = (blog) => {
-    dispatch(likeBlog(blog))
-    dispatch(
-      newNotification(`You liked blog "${blog.title}" by ${blog.author}`, 5)
-    )
-  }
-
-  const deleteBlog = (blog) => {
-    const id = blog.id
-    const blogToRemove = blogs.find((b) => b.id === id)
-
-    if (
-      window.confirm(
-        `Remove blog "${blogToRemove.title}" by ${blogToRemove.author}?`
-      )
-    ) {
-      dispatch(removeBlog(blog))
-      dispatch(newNotification(`Blog "${blogToRemove.title}" was removed`, 5))
-    }
-  }
 
   return (
     <div className="bloglist">
       {blogs.map((blog) => (
-        <Blog
-          key={blog.id}
-          blog={blog}
-          updateLikes={addLike}
-          removeBlog={deleteBlog}
-          user={user}
-        />
+        <div key={blog.id}>
+          <Link to={`/blogs/${blog.id}`}>
+            {blog.title} by {blog.author}
+          </Link>
+        </div>
       ))}
     </div>
   )
